+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Commands/Commands.cs
@@ -1,4 +1,5 @@
 #if UNITY_EDITOR && UNITY_INPUT_SYSTEM_UI_TK_ASSET_EDITOR
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
@@ -37,7 +38,12 @@ namespace UnityEngine.InputSystem.Editor
         {
             return (in InputActionsEditorState state) =>
             {
-                var actionMap = Selectors.GetSelectedActionMap(state).wrappedProperty;
+                var actionMap = Selectors.GetSelectedActionMap(state)?.wrappedProperty;
+                if (actionMap == null)
+                {
+                    Debug.LogError("Cannot add action without an action map selected");
+                    return state;
+                }
                 var newAction = InputActionSerializationHelpers.AddAction(actionMap);
                 InputActionSerializationHelpers.AddBinding(newAction, actionMap);
                 state.serializedObject.ApplyModifiedProperties();
@@ -49,13 +55,89 @@ namespace UnityEngine.InputSystem.Editor
         {
             return (in InputActionsEditorState state) =>
             {
-                var action = Selectors.GetSelectedAction(state).wrappedProperty;
-                var map = Selectors.GetSelectedActionMap(state).wrappedProperty;
+                var action = Selectors.GetSelectedAction(state)?.wrappedProperty;
+                var map = Selectors.GetSelectedActionMap(state)?.wrappedProperty;
+                if (action == null || map == null)
+                {
+                    Debug.LogError("Cannot add binding without an action and action map selected");
+                    return state;
+                }
                 var binding = InputActionSerializationHelpers.AddBinding(action, map);
                 var bindingIndex = new SerializedInputBinding(binding).indexOfBinding;
                 state.serializedObject.ApplyModifiedProperties();
                 return state.With(selectedBindingIndex: bindingIndex, selectionType: SelectionType.Binding);
             };
+        }
+
+        public static Command DeleteActionMap(int actionMapIndex)
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
+                var actionMapID = InputActionSerializationHelpers.GetId(actionMap);
+                InputActionSerializationHelpers.DeleteActionMap(state.serializedObject, actionMapID);
+                state.serializedObject.ApplyModifiedProperties();
+                if (state.selectedActionMapIndex == actionMapIndex)
+                    return SelectPrevActionMap(state);
+                return state.SelectActionMap(state.selectedActionMapIndex > actionMapIndex ? state.selectedActionMapIndex - 1 : state.selectedActionMapIndex);
+            };
+        }
+
+        private static InputActionsEditorState SelectPrevActionMap(InputActionsEditorState state)
+        {
+            var count = Selectors.GetActionMapCount(state);
+            int index = -1;
+            if (count != null && count.Value > 0)
+                index = Math.Max(state.selectedActionMapIndex - 1, 0);
+            return state.SelectActionMap(index);
+        }
+
+        public static Command DeleteAction(int actionMapIndex, string actionName)
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
+                var action = Selectors.GetActionInMap(state, actionMapIndex, actionName).wrappedProperty;
+                var actionIndex = action.GetIndexOfArrayElement();
+                var actionID = InputActionSerializationHelpers.GetId(action);
+                InputActionSerializationHelpers.DeleteActionAndBindings(actionMap, actionID);
+                state.serializedObject.ApplyModifiedProperties();
+                if (state.selectedActionIndex >= actionIndex)
+                    return SelectPrevAction(state, actionMap);
+                return state.SelectAction(state.selectedActionIndex);
+            };
+        }
+
+        private static InputActionsEditorState SelectPrevAction(InputActionsEditorState state, SerializedProperty actionMap)
+        {
+            var count = Selectors.GetActionCount(actionMap);
+            int index = -1;
+            if (count != null && count.Value > 0)
+                index = Math.Max(state.selectedActionIndex - 1, 0);
+            return state.SelectAction(index);
+        }
+
+        public static Command DeleteBinding(int actionMapIndex, int bindingIndex)
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
+                var binding = Selectors.GetCompositeOrBindingInMap(actionMap, bindingIndex).wrappedProperty;
+                InputActionSerializationHelpers.DeleteBinding(binding, actionMap);
+                state.serializedObject.ApplyModifiedProperties();
+                if (state.selectedBindingIndex >= bindingIndex)
+                    return SelectPrevBinding(state, actionMap);
+                return state.SelectBinding(state.selectedBindingIndex);
+            };
+        }
+
+        private static InputActionsEditorState SelectPrevBinding(InputActionsEditorState state, SerializedProperty actionMap)
+        {
+            var count = Selectors.GetBindingCount(actionMap);
+            var index = -1;
+            if (count != null && count.Value > 0)
+                index = Math.Max(state.selectedBindingIndex - 1, 0);
+            return state.SelectBinding(index);
         }
 
         public static Command ExpandCompositeBinding(SerializedInputBinding binding)
@@ -188,12 +270,36 @@ namespace UnityEngine.InputSystem.Editor
             };
         }
 
-        public static Command ChangeActionMapName(string newName)
+        public static Command ChangeActionMapName(int index, string newName)
         {
             return (in InputActionsEditorState state) =>
             {
-                var actionMap = Selectors.GetSelectedActionMap(state).wrappedProperty;
+                var actionMap = Selectors.GetActionMapAtIndex(state, index)?.wrappedProperty;
                 InputActionSerializationHelpers.RenameActionMap(actionMap, newName);
+                state.serializedObject.ApplyModifiedProperties();
+                return state;
+            };
+        }
+
+        public static Command ChangeActionName(int actionMapIndex, string oldName, string newName)
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
+                var action = Selectors.GetActionInMap(state, actionMapIndex, oldName).wrappedProperty;
+                InputActionSerializationHelpers.RenameAction(action, actionMap, newName);
+                state.serializedObject.ApplyModifiedProperties();
+                return state;
+            };
+        }
+
+        public static Command ChangeCompositeName(int actionMapIndex, int bindingIndex, string newName)
+        {
+            return (in InputActionsEditorState state) =>
+            {
+                var actionMap = Selectors.GetActionMapAtIndex(state, actionMapIndex)?.wrappedProperty;
+                var binding = Selectors.GetCompositeOrBindingInMap(actionMap, bindingIndex).wrappedProperty;
+                InputActionSerializationHelpers.RenameComposite(binding, newName);
                 state.serializedObject.ApplyModifiedProperties();
                 return state;
             };